refactor(soute): les gestes de la soute, du registre et des entrepôts sortent d'app.js
Dix-sept fonctions et trois clés de stockage rejoignent `soute-actions.js` :
les trois stores — la soute, le registre des chargements, les entrepôts — et
tous les gestes qui les touchent : charger une jambe, vendre ici, déposer,
reprendre, déclarer à bord, poser sa position.

Elles vivaient dans `app.js` faute d'y pouvoir sortir : chacune enchaîne une
mutation d'état, une persistance, un message et un rendu. C'est le crochet de
`rendu.ts` qui lève le dernier obstacle — elles appellent `rafraichir()`.

Déplacement PUR : aucune logique ne change, et la séparation des trois stores
reste le contrat qu'elle était — la soute se vide par plusieurs chemins dont
aucun ne rend rien à la station, donc aucun ne décharge la jambe.

Deux ajustements mécaniques seulement : les lectures de champ passent par une
aide locale au lieu du `$()` d'`app.js`, et `poserPosition` garde son nœud
absent au lieu de le déréférencer.

app.js : 1 989 → 1 834 lignes.
Vérification : 244 e2e — 242 passés, 2 ignorés, 0 échec — · 490/490 unitaires ·
tsc silencieux · build OK.
